pyre-interpreter: restore _random to PyPy parity; relocate macro smoke probes to cfg(test)
_random exposed non-PyPy public API (_unwrap_alias_probe, _PROBE_CONST,
_ProbeError, _seed_from_path, _path_bytes module attrs; __reduce__ and
raw_state on Random) that existed only to exercise the #[pyre_*] /
py_module! macro arms. Restore the module to match
pypy/module/_random/moduledef.py (interpleveldefs = { Random }), keeping
W_Random with __init__/seed/random/getrandbits/getstate/setstate.

Move the macro-arm smoke coverage into a #[cfg(test)] macro_smoke module
(W_Demo #[pyre_class] with getter/setter/deleter/__reduce__/base=, the
PyPath/text/int unwrap-alias and Vec-return #[pyre_function] probes, and a
py_module! exercising int_constants/exceptions/appleveldefs/inline_app).
Compiling the test build expands every arm; three #[test]s run the
wrappers that need no module-init bootstrap. app_random.py becomes the
test-only app_smoke.py.
